expand tests to include multiple configurations, more assertions, and optimized wait times
This pr adds a few features. First, the wait times for async actions, like
installing the operator, are minimized by replacing estimated sleeps with
polling. Particularly, the wait_for_resource and curl_retry use polling with a
few second delay to effectively wait for pods and other resources to become
available and for servers to begin accepting requests, respectively. Second,
the pr adds stricter and more test assertions. These include actions like
testing for existance (and lack thereof) of pods, testing network
accessibility, and testing for readiness by a timeout (failing to be ready by a
timeout likely indicates an error). Lastly, the pr expands the testing profile
and facilitates future expansions by generating n pipelines and creating n k3s
clusters, where n is the number of operator tests or configurations which need
a clean machine. It does so by tagging the ec2 with the test name and including
that tag in the pipeline; an extension of the same pattern used to generate and
link just the one ec2.